arrow_avro

Module codec

Data type conversions between Avro and Arrow types

This module contains the necessary types and functions to convert between Avro data types and Arrow data types.

Structs

AvroDataType

An Avro datatype mapped to the arrow data model

AvroField

A named AvroDataType

AvroFieldBuilder

Builder for an AvroField

EnumMapping

Holds the mapping information for resolving Avro enums.

Maker 🔒

Resolves Avro type names to AvroDataType

ResolvedRecord

Contains the necessary information to resolve a writer’s record against a reader’s record schema.

Resolver 🔒

Resolves Avro type names to AvroDataType

Enums

AvroLiteral 🔒

Represents a literal Avro value.

Codec

An Avro encoding

Nullability

Avro types are not nullable, with nullability instead encoded as a union where one of the variants is the null type.

Promotion 🔒

Defines the type of promotion to be applied during schema resolution.

ResolutionInfo 🔒

Contains information about how to resolve differences between a writer’s and a reader’s schema.

Functions

parse_decimal_attributes 🔒

with_extension_type 🔒