arrow_avro::reader

Module async_reader

Asynchronous implementation of Avro file reader.

This module provides AsyncAvroFileReader, which supports reading and decoding the Avro OCF format from any source that implements AsyncFileReader.

Modules

async_file_reader 🔒

builder 🔒

store 🔒

Structs

AsyncAvroFileReader

An asynchronous Avro file reader that implements Stream<Item = Result<RecordBatch, ArrowError>>. This uses an AsyncFileReader to fetch data ranges as needed, starting with fetching the header, then reading all the blocks in the provided range where:

AvroObjectReader

An implementation of an AsyncFileReader using the [ObjectStore] API.

ReaderBuilder

Builder for an asynchronous Avro file reader.

Enums

FetchNextBehaviour 🔒

ReaderState 🔒

Traits

AsyncFileReader

The asynchronous interface used by super::AsyncAvroFileReader to read avro files

Functions

read_header_info

Reads the Avro file header (magic, metadata, sync marker) asynchronously from reader.