arrow_avro::reader::header

Struct Header

pub struct Header {
    meta_offsets: Vec<usize>,
    meta_buf: Vec<u8>,
    sync: [u8; 16],
}

A decoded header for an Object Container File

Fields
meta_offsets: Vec<usize>meta_buf: Vec<u8>sync: [u8; 16]
Implementations

impl Header

pub fn metadata(&self) -> impl Iterator<Item = (&[u8], &[u8])>

Returns an iterator over the meta keys in this header

pub fn get(&self, key: impl AsRef<[u8]>) -> Option<&[u8]>

Returns the value for a given metadata key if present

pub fn sync(&self) -> [u8; 16]

Returns the sync token for this file

pub fn compression(&self) -> Result<Option<CompressionCodec>, ArrowError>

Returns the CompressionCodec if any

pub fn schema(&self) -> Result<Option<Schema<'_>>, ArrowError>

Returns the Schema if any