arrow_avro::reader::record

Struct RecordDecoder

pub(crate) struct RecordDecoder {
    schema: SchemaRef,
    fields: Vec<Decoder>,
    projector: Option<Projector>,
}

Decodes avro encoded data into [RecordBatch]

Fields
schema: SchemaReffields: Vec<Decoder>projector: Option<Projector>
Implementations

impl RecordDecoder

pub(crate) fn try_new_with_options( data_type: &AvroDataType, ) -> Result<Self, AvroError>

Creates a new RecordDecoder from the provided AvroDataType with additional options.

This method allows you to customize how the Avro data is decoded into Arrow arrays.

Arguments

data_type - The Avro data type to decode.

use_utf8view - A flag indicating whether to use Utf8View for string types.

Errors

This function will return an error if the provided data_type is not a Record.

pub(crate) fn schema(&self) -> &SchemaRef

Returns the decoder’s SchemaRef

pub(crate) fn decode( &mut self, buf: &[u8], count: usize, ) -> Result<usize, AvroError>

Decode count records from buf

pub(crate) fn flush(&mut self) -> Result<RecordBatch, AvroError>

Flush the decoded records into a [RecordBatch]