arrow_avro::reader

Struct Decoder

pub struct Decoder {
    active_decoder: RecordDecoder,
    active_fingerprint: Option<Fingerprint>,
    batch_size: usize,
    remaining_capacity: usize,
    cache: IndexMap<Fingerprint, RecordDecoder>,
    fingerprint_algorithm: FingerprintAlgorithm,
    utf8_view: bool,
    strict_mode: bool,
    pending_schema: Option<(Fingerprint, RecordDecoder)>,
    awaiting_body: bool,
}

A low-level interface for decoding Avro-encoded bytes into Arrow RecordBatch.

Fields
active_decoder: RecordDecoderactive_fingerprint: Option<Fingerprint>batch_size: usizeremaining_capacity: usizecache: IndexMap<Fingerprint, RecordDecoder>fingerprint_algorithm: FingerprintAlgorithmutf8_view: boolstrict_mode: boolpending_schema: Option<(Fingerprint, RecordDecoder)>awaiting_body: bool
Implementations

impl Decoder

pub fn schema(&self) -> SchemaRef

Return the Arrow schema for the rows decoded by this decoder

pub fn batch_size(&self) -> usize

Return the configured maximum number of rows per batch

pub fn decode(&mut self, data: &[u8]) -> Result<usize, ArrowError>

Feed data into the decoder row by row until we either:

consume all bytes in data, or

reach batch_size decoded rows.

Returns the number of bytes consumed.

fn handle_prefix(&mut self, buf: &[u8]) -> Result<Option<usize>, ArrowError>

fn handle_prefix_common<const MAGIC_LEN: usize, const N: usize>( &mut self, buf: &[u8], magic: &[u8; MAGIC_LEN], fingerprint_from: impl FnOnce([u8; N]) -> Fingerprint, ) -> Result<Option<usize>, ArrowError>

This method checks for the provided magic bytes at the start of buf and, if present, attempts to read the following fingerprint of N bytes, converting it to a Fingerprint using fingerprint_from.

fn handle_fingerprint<const N: usize>( &mut self, buf: &[u8], fingerprint_from: impl FnOnce([u8; N]) -> Fingerprint, ) -> Result<Option<usize>, ArrowError>

fn apply_pending_schema(&mut self)

fn apply_pending_schema_if_batch_empty(&mut self)

fn flush_and_reset(&mut self) -> Result<Option<RecordBatch>, ArrowError>

pub fn flush(&mut self) -> Result<Option<RecordBatch>, ArrowError>

Produce a RecordBatch if at least one row is fully decoded, returning Ok(None) if no new rows are available.

pub fn capacity(&self) -> usize

Returns the number of rows that can be added to this decoder before it is full.

pub fn batch_is_full(&self) -> bool

Returns true if the decoder has reached its capacity for the current batch.

pub fn batch_is_empty(&self) -> bool

Returns true if the decoder has not decoded any batches yet.

fn decode_block( &mut self, data: &[u8], count: usize, ) -> Result<(usize, usize), ArrowError>

fn flush_block(&mut self) -> Result<Option<RecordBatch>, ArrowError>