arrow_avro::reader

Struct ReaderBuilder

pub struct ReaderBuilder {
    batch_size: usize,
    strict_mode: bool,
    utf8_view: bool,
    schema: Option<Schema<'static>>,
}

A builder to create an Avro Reader that reads Avro data into Arrow RecordBatch.

Fields
batch_size: usizestrict_mode: boolutf8_view: boolschema: Option<Schema<'static>>
Implementations

impl ReaderBuilder

pub fn new() -> Self

Creates a new ReaderBuilder with default settings:

batch_size = 1024

strict_mode = false

utf8_view = false

schema = None

fn make_record_decoder( &self, schema: &AvroSchema<'_>, ) -> Result<RecordDecoder, ArrowError>

fn build_impl<R: BufRead>( self, reader: &mut R, ) -> Result<(Header, Decoder), ArrowError>

pub fn with_batch_size(self, batch_size: usize) -> Self

Sets the row-based batch size

pub fn with_utf8_view(self, utf8_view: bool) -> Self

Set whether to use StringViewArray for string data

When enabled, string data from Avro files will be loaded into Arrow’s StringViewArray instead of the standard StringArray.

pub fn use_utf8view(&self) -> bool

Get whether StringViewArray is enabled for string data

pub fn with_strict_mode(self, strict_mode: bool) -> Self

Controls whether certain Avro unions of the form [T, "null"] should produce an error.

pub fn with_schema(self, schema: AvroSchema<'static>) -> Self

Sets the Avro schema.

If a schema is not provided, the schema will be read from the Avro file header.

pub fn build<R: BufRead>(self, reader: R) -> Result<Reader<R>, ArrowError>

Create a Reader from this builder and a BufRead

pub fn build_decoder<R: BufRead>(self, reader: R) -> Result<Decoder, ArrowError>

Create a Decoder from this builder and a BufRead by reading and parsing the Avro file’s header. This will not create a full Reader.