arrow_avro::reader

Trait AsyncFileReader

pub trait AsyncFileReader: Send {
    // Required method
    fn get_bytes(
        &mut self,
        range: Range<u64>,
    ) -> BoxFuture<'_, Result<Bytes, AvroError>>;

    // Provided method
    fn get_byte_ranges(
        &mut self,
        ranges: Vec<Range<u64>>,
    ) -> BoxFuture<'_, Result<Vec<Bytes>, AvroError>> { ... }
}

The asynchronous interface used by super::AsyncAvroFileReader to read avro files

Notes:

There is a default implementation for types that implement [AsyncRead] and [AsyncSeek], for example tokio::fs::File.

super::AvroObjectReader, available when the object_store crate feature is enabled, implements this interface for ObjectStore.

Required Methods

fn get_bytes( &mut self, range: Range<u64>, ) -> BoxFuture<'_, Result<Bytes, AvroError>>

Retrieve the bytes in range

Provided Methods

fn get_byte_ranges( &mut self, ranges: Vec<Range<u64>>, ) -> BoxFuture<'_, Result<Vec<Bytes>, AvroError>>

Retrieve multiple byte ranges. The default implementation will call get_bytes sequentially