arrow_avro::schema

Function generate_canonical_form

pub fn generate_canonical_form(
    schema: &Schema<'_>,
) -> Result<String, ArrowError>

Generates the Parsed Canonical Form for the given Schema.

The canonical form is a standardized JSON representation of the schema, primarily used for generating a schema fingerprint for equality checking.

This form strips attributes that do not affect the schema’s identity, such as doc fields, aliases, and any properties not defined in the Avro specification.

https://avro.apache.org/docs/1.11.1/specification/#parsing-canonical-form-for-schemas