Crate arrow_avro

Convert data to / from the Apache Arrow memory format and Apache Avro

Modules

codec

Data type conversions between Avro and Arrow types

compression

Compression codec implementations for Avro

reader

Core functionality for reading Avro data into Arrow arrays

schema

Avro schema parsing and representation

writer

Core functionality for writing Arrow arrays as Avro data

Traits

AvroFieldExt

Extension trait for AvroField to add Utf8View support