arrow_avro

Module codec

Data type conversions between Avro and Arrow types

This module contains the necessary types and functions to convert between Avro data types and Arrow data types.

Structs

AvroDataType

An Avro datatype mapped to the arrow data model

AvroField

A named AvroDataType

AvroFieldBuilder

Builder for an AvroField

Resolver 🔒

Resolves Avro type names to AvroDataType

Enums

Codec

An Avro encoding

Nullability

Avro types are not nullable, with nullability instead encoded as a union where one of the variants is the null type.

Functions

make_data_type 🔒

Parses a AvroDataType from the provided Schema and the given name and namespace

parse_decimal_attributes 🔒

with_extension_type 🔒