arrow_avro

Module codec

Structs

AvroDataType

An Avro datatype mapped to the arrow data model

AvroField

A named AvroDataType

Resolver 🔒

Resolves Avro type names to AvroDataType

Enums

Codec

An Avro encoding

Nullability

Avro types are not nullable, with nullability instead encoded as a union where one of the variants is the null type.

Functions

make_data_type 🔒

Parses a AvroDataType from the provided Schema and the given name and namespace