arrow_avro::reader

Module async_reader

Re-exports

pub use async_file_reader::AsyncFileReader;

pub use builder::ReaderBuilder;

pub use store::AvroObjectReader;

Modules

async_file_reader 🔒

builder 🔒

store 🔒

Structs

AsyncAvroFileReader

An asynchronous Avro file reader that implements Stream<Item = Result<RecordBatch, ArrowError>>. This uses an AsyncFileReader to fetch data ranges as needed, starting with fetching the header, then reading all the blocks in the provided range where:

Enums

FetchNextBehaviour 🔒

ReaderState 🔒