Module arrow_avro::schema

Structs

Array

An array

Attributes

Additional attributes within a Schema

Enum

An enumeration

Field

A field within a Record

Fixed

A fixed length binary array

Map

A map

Record

A record

Type

A type definition that is not a variant of ComplexType

Enums

ComplexType

A complex type

PrimitiveType

A primitive type

Schema

An Avro schema

TypeName

Either a PrimitiveType or a reference to a previously defined named type

Constants

SCHEMA_METADATA_KEY

The metadata key used for storing the JSON encoded Schema